requirements.txt
moviepy<2.0
pymupdf
imageio-ffmpeg
fastapi
uvicorn
python-multipart
Pillow
requests
pytest
cutlet
unidic-lite
coqui-tts
sounddevice
soundfile
scipy
numpy
torch
torchaudio
